Refactor: add method to check acct card params; rename helpers
Helper methods renamed: these are for handling contact info, if present in transaction params. They should not interfere with params that have no contact info in them, as those can still be valid transactions. It may be beneficial to do further refactoring of transaction creation logic to make this flow easier to follow.